Added reservation valid checks, fixed a lot of bs
